changed to version 0.1.0 changed the server port to 8075 changed message object to have a sender and a receiver changed MessageController to use simpMessageTemplate to send messages to the right receiver and also to the sender to send the message to all active devices for the same user changed the online chat client example to use a own channel to receive messages

+++ src/main/java/de/judgeman/messenger/repository/MessageRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 /**
  * Created by Paul Richter on Sat 24/07/2021
  */
-public interface MessageRepository extends CrudRepository<Message, Integer> {
+public interface MessageRepository extends CrudRepository<Message, Long> {
 
     List<Message> findAll();
 }
